eegcpm-0.1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eegcpm"
version = "0.1.0"
description = "EEG Connectome Predictive Modeling Toolbox"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "EEGCPM Team"}
]
keywords = ["EEG", "connectivity", "predictive modeling", "neuroscience", "MNE"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "pandas>=1.3",
    "mne>=1.0",
    "scikit-learn>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pyzaplineplus>=0.1.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.20",
    "plotly>=5.0",
]
neural = [
    "torch>=2.0",
]
all = [
    "streamlit>=1.20",
    "plotly>=5.0",
    "torch>=2.0",
    "autoreject>=0.4",
    "nibabel>=4.0",
    "networkx>=2.8",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
eegcpm = "eegcpm.cli:main"

[project.urls]
Homepage = "https://github.com/eegcpm/eegcpm"
Documentation = "https://eegcpm.readthedocs.io"
Repository = "https://github.com/eegcpm/eegcpm"

[tool.setuptools.packages.find]
where = ["."]
include = ["eegcpm*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=eegcpm --cov-report=term-missing"
